Unified the payments in directDebit so that only a transaction is made with payments pending members and programs. It became mandatory to inform the paying user when registering or modifying a program.
It added the form of payment with bank income which counts of 3 buttons
that is to mark, to cancel and to accept payment.
 - The user can only dial and cancel once.
 - The administrator can dial, cancel as many times as he wants, but
accept payment once.
 - The administrator can also cancel a payment in case it is wrong.

In the view of payments there is a refresh payments if this something
wrong is recharged, that is to say they are deleted or are re-saved if
necessary.

Bank remittances take directDebit data instead of directly uploading it.
It was added angular to the entire payment system.
Changed the list of users generated with thymeleaf by the one of
angular.

A modal angularjs was added when you clicked on the user's name to
quickly see all your information.
An alert is issued when making any changes to one of the angular tables.
 - Angular tables have sorting, searching and pagination.
The sping.io library has been updated to version 2.0.8 to keep the
project updated.