0.2.5 - managements, events, payments and bankRemitances XML, firebase Perfection of everything related to Firebase  - Authentication  - Registration (Only Firebase, members or both)  - Reset password sending email  - Restore Password  - Validation of tokens
Accounts:
 - New EXUSER user, who is a user that can access members but can not
participate in activities or programs.
 - The administrator can see the payments of the partners, in the
profile view of that user and if he does not have to create a new one.
 - From the event view the administrator can access the user profile.
 - The user can request to be registered and low.

In the configuration view has been added the possibility of loading
users and programs using a JSON.
 - The users, a new one is created or it is actualized in the function
of the mail.
 - The programs, a new one is created or it is actualized in the mail
function, also it will be created the members if there is relation.

Other small corrections and improvements.
Clean code and formatting with the same style